Extract base class to separate file
